Initial fix for file uploads on Android Honeycomb (tested Samsung Galaxy Tab 10.1)
Android browser wasn't supporting sendAsBinary, nor was it supporting the binary string funkiness we used. Luckily it *does* support using the blob directly, and that seems to work in other browsers as well so far. Yay!
The Firefox label click hack unfortunately interferes with Android -- it triggers the file selection but then fails to send the 'change' event, so we never picked it up.

Main gotcha remaining is that we don't get a 'type' attribute on the blob; it's there but empty.
For now I'm "faking it" by sending 'image/jpeg' if we have no type, but this will fail at some point.
We'll probably need server-side file type detection.